Change: Move release argument parser into own module
Separate the parsing from running the CLI functions.
